Improve error handling
We now use `MonadBaseControl` (/ base-lifted) so that worker actions run in a user specified monad. Moreover, monitoring now takes a callback that will be notified on worker status updates, so that the caller could e.g. react to a worker crash.

Also somewhere along the way, I renamed BotM => Replicant. Hope that's okay.